CI_CD/requirements.txt
tensorflow==2.10.0
torch==1.13.1
torchvision==0.14.1
streamlit==1.19.0
kafka-python==2.0.2
Pillow==9.4.0
requests==2.28.2
GitPython==3.1.30
apache-airflow==2.5.1
